new utils dir cause old one had model dependencies
