+++ escalate/core/utils_no_dependencies/view_utils.py
@@ -1,23 +1,4 @@
-import functools
-import random
-import string
-import copy
-import json
-import collections.abc
-
-
-def rsetattr(obj, attr, val):
-    pre, _, post = attr.rpartition('.')
-    return setattr(rgetattr(obj, pre) if pre else obj, post, val)
-
-
-def rgetattr(obj, attr, *args):
-    def _getattr(obj, attr):
-        if obj == None:
-            return None
-        else:
-            return getattr(obj, attr, *args)
-    return functools.reduce(_getattr, [obj] + attr.split('.'))
+from .misc import FrozenDict
 
 def get_all_related_fields_backward(model):
     def _get_all_fields_rec(model, all_related_fields, parent_model=None, parent_field_name=None):
@@ -104,46 +85,3 @@ def get_model_of_related_field(model, full_related_field, all_related_fields=Non
         })
         cur_model = all_related_fields[cur_field_details]
     return cur_field_details['model']
-
-
-class FrozenDict(collections.abc.Mapping):
-    '''
-    Frozen dictionary class to use if needed
-    item assignment is disabled
-    new attribute assignment is disabled since it doesn't extend dict directly
-    '''
-    __is_frozen = False
-
-    def __init__(self, *args, **kwargs):
-        self._d = dict(*args, **kwargs)
-        self._hash = None
-        self.__is_frozen = True
-
-    def __iter__(self):
-        return iter(self._d)
-
-    def __len__(self):
-        return len(self._d)
-
-    def __getitem__(self, key):
-        return self._d[key]
-
-    def __str__(self):
-        return str(self._d)
-
-    def __repr__(self):
-        return str(self._d)
-
-    def __hash__(self):
-        if self._hash is None:
-            hash_ = 0
-            for pair in sorted(self.items()):
-                hash_ ^= hash(pair)
-            self._hash = hash_
-        return self._hash
-
-    def __setattr__(self, key, value):
-        if self.__is_frozen and not hasattr(self, key):
-            raise Exception(
-                f'{self} is a frozen class. Cannot set new attribute {key}')
-        object.__setattr__(self, key, value)
